+++ src/Bazar.php
@@ -17,7 +17,7 @@ abstract class Bazar
      *
      * @var string
      */
-    public const VERSION = '0.2.0';
+    public const VERSION = '0.2.1';
 
     /**
      * The default currency.